# api/validators/message.ts
import { z } from 'zod';

export const MessageRequestSchema = z.object({
  target: z.string().min(1, 'Target is required'),
  message: z.string().min(1, 'Message is required'),
  execute: z.boolean().optional().default(false)
});

export type MessageRequest = z.infer<typeof MessageRequestSchema>;

export const MessageResponseSchema = z.object({
  status: z.enum(['acknowledged', 'failed']),
  timestamp: z.string(),
  target: z.string(),
  message: z.string().optional(),
  error: z.string().optional()
});

export type MessageResponse = z.infer<typeof MessageResponseSchema>;