# openapi.yaml
openapi: 3.0.3
info:
  title: MAS (Multi-Agent System) API
  description: |
    MAS（マルチエージェントシステム）を管理・実行するためのHTTP APIです。
    このAPIを使用することで、複数のAIエージェントを並列で動作させる構成を
    WebUIから送信し、tmuxセッション内でシステムを起動できます。

    ## 主要機能
    - MASセッションの起動と管理（/runs）
    - tmuxセッションへのメッセージ送信（/message）

    ## 設計原則
    - APIは「起動のトリガー」として機能
    - 状態管理はtmuxとファイルシステムに委譲
    - UUIDベースのセッション識別
    - ステートレスな実装
  version: 2.0.0
  contact:
    name: MAS Support Team
  license:
    name: MIT

servers:
  - url: http://localhost:8765
    description: ローカル開発環境（デフォルト）
  - url: http://{host}:{port}
    description: カスタムサーバー
    variables:
      host:
        default: 0.0.0.0
        description: サーバーのホストアドレス
      port:
        default: '8765'
        description: サーバーのポート番号

tags:
  - name: Runs
    description: MASセッションの実行管理
  - name: Message
    description: tmuxセッションへのメッセージ送信

paths:
  /runs:
    post:
      summary: MASセッションを開始
      description: |
        エージェント構成を受け取り、新しいMASセッションを開始します。
        このエンドポイントは以下の処理を実行します：

        1. UUIDを生成してセッションIDとする
        2. セッション用ディレクトリを作成
        3. 設定を`config.json`として保存
        4. tmuxセッションを起動
        5. 各エージェントをtmux上で起動

        **注意**: APIは起動の完了を待たずに即座にレスポンスを返します。
        実行状態の確認はtmuxまたはファイルシステムを直接参照してください。
      tags:
        - Runs
      operationId: createRun
      requestBody:
        description: MASエージェント構成
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
            examples:
              singleUnit:
                summary: 単一ユニット構成（メタマネージャーなし）
                value:
                  agents:
                    units:
                      - unitId: 1
                        manager:
                          id: "10"
                          prompt: "このunitの進行管理と判断を担当する"
                        workers:
                          - id: "11"
                            prompt: "設計を担当する"
                          - id: "12"
                            prompt: "実装を担当する"
                          - id: "13"
                            prompt: "テストを担当する"
              multiUnit:
                summary: 複数ユニット構成（メタマネージャー付き）
                value:
                  agents:
                    metaManager:
                      id: "00"
                      prompt: "全体を統合し、各unitの成果をまとめて意思決定を行う"
                    units:
                      - unitId: 1
                        manager:
                          id: "10"
                          prompt: "設計・開発unitの管理"
                        workers:
                          - id: "11"
                            prompt: "アーキテクチャ設計"
                          - id: "12"
                            prompt: "バックエンド実装"
                          - id: "13"
                            prompt: "フロントエンド実装"
                      - unitId: 2
                        manager:
                          id: "20"
                          prompt: "品質保証unitの管理"
                        workers:
                          - id: "21"
                            prompt: "単体テスト作成"
                          - id: "22"
                            prompt: "統合テスト実施"
                          - id: "23"
                            prompt: "パフォーマンステスト"
              fullScale:
                summary: フルスケール構成（4ユニット）
                value:
                  agents:
                    metaManager:
                      id: "00"
                      prompt: "プロジェクト全体の統括と意思決定"
                    units:
                      - unitId: 1
                        manager:
                          id: "10"
                          prompt: "設計unitの管理"
                        workers:
                          - id: "11"
                            prompt: "要件分析"
                          - id: "12"
                            prompt: "アーキテクチャ設計"
                          - id: "13"
                            prompt: "データベース設計"
                      - unitId: 2
                        manager:
                          id: "20"
                          prompt: "開発unitの管理"
                        workers:
                          - id: "21"
                            prompt: "バックエンド開発"
                          - id: "22"
                            prompt: "フロントエンド開発"
                          - id: "23"
                            prompt: "API開発"
                      - unitId: 3
                        manager:
                          id: "30"
                          prompt: "品質unitの管理"
                        workers:
                          - id: "31"
                            prompt: "テスト戦略策定"
                          - id: "32"
                            prompt: "自動テスト実装"
                          - id: "33"
                            prompt: "品質メトリクス分析"
                      - unitId: 4
                        manager:
                          id: "40"
                          prompt: "運用unitの管理"
                        workers:
                          - id: "41"
                            prompt: "デプロイメント"
                          - id: "42"
                            prompt: "監視設定"
                          - id: "43"
                            prompt: "ドキュメント作成"

      responses:
        '201':
          description: MASセッションが正常に開始された
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
              example:
                sessionId: "6f9c2e5a-8f8a-4a6d-b3f1-9c9e0c8a7e21"
                tmuxSession: "6f9c2e5a-8f8a-4a6d-b3f1-9c9e0c8a7e21"
                workingDir: "sessions/6f9c2e5a-8f8a-4a6d-b3f1-9c9e0c8a7e21"
                startedAt: "2024-12-15T10:30:00.000Z"

        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingAgents:
                  summary: agents フィールドが欠けている
                  value:
                    error: "agents field is required"
                    code: "MISSING_FIELD"
                emptyUnits:
                  summary: units配列が空
                  value:
                    error: "units array cannot be empty"
                    code: "INVALID_STRUCTURE"
                invalidJson:
                  summary: JSONフォーマットが不正
                  value:
                    error: "Invalid JSON format"
                    code: "PARSE_ERROR"

        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to create session directory"
                code: "INTERNAL_ERROR"

      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X POST http://localhost:8765/runs \
              -H "Content-Type: application/json" \
              -d '{
                "agents": {
                  "metaManager": {
                    "id": "00",
                    "prompt": "全体統括"
                  },
                  "units": [
                    {
                      "unitId": 1,
                      "manager": {
                        "id": "10",
                        "prompt": "unit1管理"
                      },
                      "workers": [
                        {"id": "11", "prompt": "設計"},
                        {"id": "12", "prompt": "実装"},
                        {"id": "13", "prompt": "テスト"}
                      ]
                    }
                  ]
                }
              }'

        - lang: JavaScript
          label: JavaScript (fetch)
          source: |
            const response = await fetch('http://localhost:8765/runs', {
              method: 'POST',
              headers: {
                'Content-Type': 'application/json',
              },
              body: JSON.stringify({
                agents: {
                  metaManager: {
                    id: '00',
                    prompt: '全体統括'
                  },
                  units: [
                    {
                      unitId: 1,
                      manager: {
                        id: '10',
                        prompt: 'unit1管理'
                      },
                      workers: [
                        { id: '11', prompt: '設計' },
                        { id: '12', prompt: '実装' },
                        { id: '13', prompt: 'テスト' }
                      ]
                    }
                  ]
                }
              })
            });

            const data = await response.json();
            console.log('Session started:', data.sessionId);

        - lang: Python
          label: Python (requests)
          source: |
            import requests
            import json

            config = {
                "agents": {
                    "metaManager": {
                        "id": "00",
                        "prompt": "全体統括"
                    },
                    "units": [
                        {
                            "unitId": 1,
                            "manager": {
                                "id": "10",
                                "prompt": "unit1管理"
                            },
                            "workers": [
                                {"id": "11", "prompt": "設計"},
                                {"id": "12", "prompt": "実装"},
                                {"id": "13", "prompt": "テスト"}
                            ]
                        }
                    ]
                }
            }

            response = requests.post(
                'http://localhost:8765/runs',
                json=config
            )

            if response.status_code == 201:
                result = response.json()
                print(f"Session ID: {result['sessionId']}")
                print(f"tmux attach -t {result['tmuxSession']}")

    options:
      summary: CORS プリフライトリクエスト
      description: ブラウザからのクロスオリジンリクエストのためのプリフライト
      tags:
        - Runs
      responses:
        '200':
          description: CORS設定を返す
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Access-Control-Allow-Methods:
              schema:
                type: string
                example: POST, OPTIONS
            Access-Control-Allow-Headers:
              schema:
                type: string
                example: Content-Type

  /message:
    post:
      summary: tmuxセッションにメッセージを送信
      description: |
        指定されたtmuxウィンドウまたはペインにメッセージを送信します。
        executeフラグを使用することで、メッセージをコマンドとして実行することも可能です。
        MASセッション起動後、特定のエージェントウィンドウに追加の指示を送る際に使用できます。
      tags:
        - Message
      operationId: sendMessage
      requestBody:
        description: 送信するメッセージの詳細
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
            examples:
              simpleMessage:
                summary: シンプルなメッセージ送信
                value:
                  target: window1
                  message: Hello, tmux!
              executeCommand:
                summary: コマンドの実行
                value:
                  target: window2
                  message: ls -la
                  execute: true
              broadcastMessage:
                summary: 全ウィンドウへのブロードキャスト
                value:
                  target: all
                  message: System maintenance in 5 minutes
              agentMessage:
                summary: 特定のエージェントへのメッセージ
                value:
                  target: agent-11
                  message: 追加のタスクを実行してください

      responses:
        '200':
          description: メッセージが正常に受理された
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                status: acknowledged
                target: window1
                timestamp: '2024-12-15T10:30:00.000Z'

        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingFields:
                  summary: 必須フィールドが欠けている
                  value:
                    error: target and message required
                    code: MISSING_FIELD
                invalidJson:
                  summary: JSONフォーマットが不正
                  value:
                    error: Unexpected token in JSON
                    code: PARSE_ERROR

      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X POST http://localhost:8765/message \
              -H "Content-Type: application/json" \
              -d '{"target": "window1", "message": "Hello, tmux!"}'

        - lang: JavaScript
          label: JavaScript (fetch)
          source: |
            fetch('http://localhost:8765/message', {
              method: 'POST',
              headers: {
                'Content-Type': 'application/json',
              },
              body: JSON.stringify({
                target: 'window1',
                message: 'Hello, tmux!'
              })
            })
            .then(response => response.json())
            .then(data => console.log(data));

        - lang: Python
          label: Python (requests)
          source: |
            import requests

            response = requests.post(
                'http://localhost:8765/message',
                json={
                    'target': 'window1',
                    'message': 'Hello, tmux!'
                }
            )
            print(response.json())

    options:
      summary: CORS プリフライトリクエスト
      description: ブラウザからのクロスオリジンリクエストのためのプリフライト
      tags:
        - Message
      responses:
        '200':
          description: CORS設定を返す
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Access-Control-Allow-Methods:
              schema:
                type: string
                example: POST, OPTIONS
            Access-Control-Allow-Headers:
              schema:
                type: string
                example: Content-Type

components:
  schemas:
    # MAS実行関連のスキーマ
    RunRequest:
      type: object
      required:
        - agents
      properties:
        agents:
          type: object
          description: エージェント構成の定義
          required:
            - units
          properties:
            metaManager:
              $ref: '#/components/schemas/Agent'
              description: |
                メタマネージャー（全体統括エージェント）。
                2ユニット以上の場合に必要です。
            units:
              type: array
              description: ユニットの配列（1～4ユニット）
              minItems: 1
              maxItems: 4
              items:
                $ref: '#/components/schemas/Unit'

    Agent:
      type: object
      description: 個別エージェントの定義
      required:
        - id
        - prompt
      properties:
        id:
          type: string
          description: |
            エージェントのID（2桁の数字文字列）。
            - "00": メタマネージャー
            - "10", "20", "30", "40": ユニットマネージャー
            - "11-13", "21-23", "31-33", "41-43": ワーカー
          pattern: '^[0-9]{2}$'
          example: "00"
        prompt:
          type: string
          description: エージェントに与える役割・指示のプロンプト
          minLength: 1
          maxLength: 5000
          example: "全体を統括し、各unitの成果をまとめて意思決定を行う"

    Unit:
      type: object
      description: ユニット（マネージャー1体 + ワーカー3体）の定義
      required:
        - unitId
        - manager
        - workers
      properties:
        unitId:
          type: integer
          description: ユニット番号（1～4）
          minimum: 1
          maximum: 4
          example: 1
        manager:
          $ref: '#/components/schemas/Agent'
          description: このユニットのマネージャーエージェント
        workers:
          type: array
          description: このユニットのワーカーエージェント（通常3体）
          minItems: 1
          maxItems: 5
          items:
            $ref: '#/components/schemas/Agent'

    RunResponse:
      type: object
      description: MAS実行開始のレスポンス
      required:
        - sessionId
        - tmuxSession
        - workingDir
        - startedAt
      properties:
        sessionId:
          type: string
          format: uuid
          description: |
            このMAS実行の一意な識別子。
            以下の用途で使用されます：
            - セッション全体の識別
            - ログやファイルの参照
            - 監視・管理
          example: "6f9c2e5a-8f8a-4a6d-b3f1-9c9e0c8a7e21"
        tmuxSession:
          type: string
          description: |
            tmuxセッション名（sessionIdと同一）。
            アタッチコマンド: `tmux attach -t {tmuxSession}`
          example: "6f9c2e5a-8f8a-4a6d-b3f1-9c9e0c8a7e21"
        workingDir:
          type: string
          description: |
            セッションのワーキングディレクトリ。
            以下のファイル・ディレクトリが含まれます：
            - config.json: エージェント構成
            - logs/: 実行ログ
            - outputs/: 成果物
          example: "sessions/6f9c2e5a-8f8a-4a6d-b3f1-9c9e0c8a7e21"
        startedAt:
          type: string
          format: date-time
          description: セッション開始時刻（ISO 8601形式）
          example: "2024-12-15T10:30:00.000Z"

    # メッセージ送信関連のスキーマ
    MessageRequest:
      type: object
      required:
        - target
        - message
      properties:
        target:
          type: string
          description: |
            メッセージの送信先。以下の形式がサポートされています：
            - ウィンドウ名（例: "window1", "agent-11"）
            - ウィンドウインデックス（例: "0", "1"）
            - ペイン指定（例: "window1.0", "0.1"）
            - セッションID.ウィンドウ名（例: "6f9c2e5a.agent-11"）
            - "all"（すべてのウィンドウ）
            - "current"（現在のウィンドウ）
          example: window1

        message:
          type: string
          description: |
            送信するメッセージまたは実行するコマンド。
            executeフラグがtrueの場合、このメッセージはシェルコマンドとして実行されます。
          example: Hello, tmux!

        execute:
          type: boolean
          description: |
            trueの場合、messageをコマンドとして実行します。
            falseまたは省略時は、テキストとして送信します。
          default: false
          example: false

    SuccessResponse:
      type: object
      required:
        - status
        - target
        - timestamp
      properties:
        status:
          type: string
          description: レスポンスステータス
          enum: [acknowledged]
          example: acknowledged

        target:
          type: string
          description: メッセージが送信された対象
          example: window1

        timestamp:
          type: string
          format: date-time
          description: メッセージを受理した時刻（ISO 8601形式）
          example: '2024-12-15T10:30:00.000Z'

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: エラーメッセージ
          example: "agents field is required"
        code:
          type: string
          description: エラーコード
          enum:
            - MISSING_FIELD
            - INVALID_STRUCTURE
            - PARSE_ERROR
            - INTERNAL_ERROR
          example: "MISSING_FIELD"
        details:
          type: object
          description: 追加のエラー詳細情報
          additionalProperties: true

  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        将来的な認証機能のためのAPIキー。
        現在は実装されていませんが、本番環境では認証の追加を推奨します。

x-webhooks:
  agentStatusUpdate:
    post:
      summary: エージェント状態更新通知（将来機能）
      description: |
        エージェントの状態が変化した際の通知。
        現在は未実装ですが、将来的なWebhook機能として予約。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
                  format: uuid
                agentId:
                  type: string
                status:
                  type: string
                  enum: [starting, running, completed, failed]
                timestamp:
                  type: string
                  format: date-time